requirements.txt
pcapy
beautifulsoup4==4.5.1
bs4==0.0.1
click==6.6
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
lxml==3.7.0
MarkupSafe==0.23
requests==2.12.4
scapy-python3==0.20
SQLAlchemy==1.1.4
Werkzeug==0.11.11
